refactor: rename packet to mse-audio-wrapper, reorg file structure
