+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "isobmff-audio",
-  "version": "0.1.1",
-  "description": "Library that wraps audio in the ISO Base Media File Format (MPEG-4 Part 12)",
+  "name": "mse-audio-wrapper",
+  "version": "1.0.0",
+  "description": "Library to enable Media Source Extensions API playback for unsupported audio containers and raw codecs",
   "main": "index.js",
   "keywords": [
     "isobmff",
@@ -9,6 +9,9 @@
     "fmp4",
     "fragmented mp4",
     "mpeg4",
+    "webm",
+    "media source extensions",
+    "mse",
     "mp3",
     "mpeg",
     "aac",
@@ -22,7 +25,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/eshaz/isobmff-audio.git"
+    "url": "git+https://github.com/eshaz/mse-audio-wrapper.git"
   },
   "author": {
     "name": "Ethan Halsall",
@@ -30,9 +33,9 @@
   },
   "license": "LGPL-3.0-or-later",
   "bugs": {
-    "url": "https://github.com/eshaz/isobmff-audio/issues"
+    "url": "https://github.com/eshaz/mse-audio-wrapper/issues"
   },
-  "homepage": "https://github.com/eshaz/isobmff-audio#readme",
+  "homepage": "https://github.com/eshaz/mse-audio-wrapper#readme",
   "devDependencies": {
     "prettier": "^2.2.1"
   }
